Define data structure for saving subscribers
In order to store emails, a simple table with an `email` column has been
provided, along with a `created` column to keep the time of creation for
potential future data analytics purposes.

For now, a unique ID should be generated and provided at runtime for the
`id` column until another, better solution replaces this that doesn't
require a runtime id generator and is instead created by the database
itself.