Provide production container for application encapsulation
In order to ensure consistent builds and runtime behavior across all
deployment environments, a `Dockerfile` has been provided. The build
stage includes all of the necessary packages to compile the application
to a distributable via GHC.

The last stage contains only the compiled binary along with `libffi` &
`gmp` which are required by the binary in order to execute.